Put test file descript in a separate helper file so we can use it for multiple tests.
